## Motivation

This PR provides a big jump in speed for WASM by leveraging SIMD instructions for `qX_K_q8_K` and `qX_0_q8_0` dot product functions.

Surprisingly, 99% of the code in this PR is **written by DeekSeek-R1**. The only thing I do is to develop tests and write prompts (with some trials and errors)

Here is an example of the prompt that I used: https://gist.github.com/ngxson/307140d24d80748bd683b396ba13be07

Indeed, this PR aims to prove that LLMs are now capable of writing good low-level code, to a point that it can optimize its own code.

## Development

To test it, I developed 2 examples using WASM and JS:
- https://github.com/ngxson/ggml/tree/xsn/wasm_simd_wip adds an example linked against `ggml.h` and `ggml-cpu.h`, used during development
- https://github.com/ngxson/wllama/pull/151 exposes an equivalent of `llama-bench` and `llama-perplexity`, used for validation and benchmark

## Result

| model | threads | test | t/s (master) | t/s (PR) | diff |
| --- | --- | --- | --- | --- | --- |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | pp 32 | 39.19 ± 0.43 | 110.37 ± 4.12 | 2.82 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | pp 64 | 39.45 ± 0.12 | 109.84 ± 2.79 | 2.78 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | pp 128 | 39.04 ± 0.04 | 106.64 ± 0.27 | 2.73 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | pp 256 | 37.73 ± 0.50 | 98.69 ± 0.15 | 2.62 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | tg 32 | 27.32 ± 0.35 | 69.98 ± 0.61 | 2.56 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | tg 64 | 26.90 ± 0.19 | 68.33 ± 2.09 | 2.54 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | tg 128 | 26.49 ± 0.72 | 66.39 ± 2.58 | 2.51 |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 7 | tg 256 | 25.56 ± 0.42 | 61.82 ± 1.02 | 2.42 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | pp 32 | 31.14 ± 0.12 | 89.03 ± 2.93 | 2.86 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | pp 64 | 31.31 ± 0.09 | 89.17 ± 0.50 | 2.85 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | pp 128 | 31.07 ± 0.11 | 86.77 ± 0.24 | 2.79 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | pp 256 | 29.76 ± 0.77 | 80.56 ± 3.23 | 2.71 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | tg 32 | 23.15 ± 0.08 | 58.42 ± 1.17 | 2.52 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | tg 64 | 22.82 ± 0.15 | 58.45 ± 0.32 | 2.56 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | tg 128 | 22.42 ± 0.42 | 57.38 ± 0.54 | 2.56 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 7 | tg 256 | 21.80 ± 0.26 | 55.59 ± 0.16 | 2.55 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | pp 32 | 42.15 ± 0.22 | 84.34 ± 2.63 | 2.00 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | pp 64 | 42.22 ± 0.39 | 84.63 ± 0.56 | 2.00 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | pp 128 | 41.89 ± 0.01 | 82.38 ± 0.37 | 1.97 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | pp 256 | 40.34 ± 0.52 | 77.53 ± 0.02 | 1.92 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | tg 32 | 28.56 ± 0.31 | 56.99 ± 0.61 | 2.00 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | tg 64 | 28.15 ± 0.15 | 56.43 ± 0.25 | 2.00 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | tg 128 | 27.67 ± 0.24 | 55.70 ± 0.34 | 2.01 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 7 | tg 256 | 26.44 ± 0.41 | 54.21 ± 0.20 | 2.05 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | pp 32 | 38.85 ± 0.24 | 71.16 ± 1.88 | 1.83 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | pp 64 | 39.01 ± 0.33 | 71.35 ± 0.32 | 1.83 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | pp 128 | 38.67 ± 0.12 | 69.65 ± 0.57 | 1.80 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | pp 256 | 37.04 ± 0.87 | 66.26 ± 0.12 | 1.79 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | tg 32 | 26.53 ± 0.24 | 51.72 ± 0.92 | 1.95 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | tg 64 | 26.25 ± 0.18 | 51.47 ± 0.41 | 1.96 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | tg 128 | 25.64 ± 0.74 | 50.80 ± 0.38 | 1.98 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 7 | tg 256 | 24.57 ± 0.03 | 49.59 ± 0.32 | 2.02 |

With perplexity mostly the same between 2 version (scalar vs SIMD):

Scalar:

| model | PPL | n_tokens |
| --- | --- | --- |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 8.599498036422444 | 2048 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 9.089770029113614 | 2048 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 8.714560764156719 | 2048 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 8.552368126397347 | 2048 |

SIMD:

| model | PPL | n_tokens |
| --- | --- | --- |
| Llama-3.2-1B-Instruct-Q8_0.gguf | 8.606497809194186 | 2048 |
| Llama-3.2-1B-Instruct-Q4_0.gguf | 9.098083009304258 | 2048 |
| Llama-3.2-1B-Instruct-Q4_K_M.gguf | 8.727644568471817 | 2048 |
| Llama-3.2-1B-Instruct-Q5_K_L.gguf | 8.552417739186847 | 2048 |