@gut5 2025-01-27
Imagine if this starts happening across the entire system/coding infrastructure. People just running critical code through Deepseek to improve it.