@gut5 2025-01-27
> > > Imagine if this starts happening across the entire system/coding infrastructure. People just running critical code through Deepseek to improve it.
> > 
> > 
> > It will happen autonomously soon.
> 
> cant wait for the first ai generated buffer overflow vuln, if it didnt already happen

of course, buffer overflows never happened with human programmers