@makedir 2025-01-27
> > > Imagine if this starts happening across the entire system/coding infrastructure. People just running critical code through Deepseek to improve it.
> > 
> > 
> > It will happen autonomously soon.
> 
> cant wait for the first ai generated buffer overflow vuln, if it didnt already happen

only humans are able to implement errors like buffer oberflow. machines are perfect after a critical point, they wont do mistakes anymore. just use good enough trained LLMs to replace humans as software developers. humans will always do mistakes.