@Kreijstal 2025-01-27
imagine if you create a github issue and github automatically writes a PR