@0x3C50 2025-01-28
> > cant wait for the first ai generated buffer overflow vuln, if it didnt already happen
> 
> 
> 
> only humans are able to implement errors like buffer oberflow. machines are perfect after a critical point, they wont do mistakes anymore. just use good enough trained LLMs to replace humans as software developers. humans will always do mistakes.

the ai has learned from code that a human or another human-trained ai has written. as long as any human-influenced code is present (there WILL only be human-influenced code to train on), the probability for error is still there. be that factual bugs like incorrect logic or more low level, system related bugs like buffer overflows. an ai can't magically stop producing incorrect code