@makedir 2025-01-28
> the ai has learned from code that a human or another human-trained ai has written.

No. obviously you learn a LLM with perfect code and then you will get always from ther perfect code. LLMs will get better the better quality the source is it is trained with. Human developers need to be replaced in the future.