@aeiouaeiouaeiouaeiouaeiouaeiou 2025-01-28
This pull request invented [Droste effect](https://en.wikipedia.org/wiki/Droste_effect) in AI.