@AlhasanIQ 2025-01-28
“Near the singularity; crystal-clear which side”