@alec-c4 2025-01-28
> Imagine if this starts happening across the entire system/coding infrastructure. People just running critical code through Deepseek to improve it.

Sorry :))))))))

![Skynet-Protocol-1024x576](https://github.com/user-attachments/assets/6e20d787-0a56-4ee7-b62d-43e6ebc471fa)