@anthogez 2025-01-28
> imagine if you create a github issue and github automatically writes a PR

@Kreijstal I’d suggest taking a look at Arvion: https://github.com/microsoftgraph/msgraph-sample-reactspa/pull/356. 
It’s already addressing gaps that tools like Dependabot miss, but within a security context.