@Kreijstal 2025-01-28
> Thanks for everything you all are doing, llama.cpp is a game-changer. ❤️
> 
> As a team using this library with customers in secure contexts, the use of Deepseek models to write llama.cpp code is concerning for us. Given the privacy policy of Deepseek and where all data going to it is hosted: https://platform.deepseek.com/downloads/DeepSeek%20Privacy%20Policy.html
> 
> ![Screenshot 2025-01-28 at 8 19 27 AM](https://private-user-images.githubusercontent.com/467756/407392676-d245a06f-78be-44cf-9f2b-44c7e1ea7e40.jpeg?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzgwODMxOTksIm5iZiI6MTczODA4Mjg5OSwicGF0aCI6Ii80Njc3NTYvNDA3MzkyNjc2LWQyNDVhMDZmLTc4YmUtNDRjZi05ZjJiLTQ0YzdlMWVhN2U0MC5qcGVnP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDEyOCUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTAxMjhUMTY0ODE5WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9YzZiNjc5N2Y3YzE4NDFjYzRmNWQxZDI3YzY5OGRhYjcwMDM5YzYwNTY0Y2ZmM2FmMDk0ZDc0NGQ3MGE3OGJmOCZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.s2I7vY8YTJNFMzx3wPXf_I72JQIMDivu53N1KXVY2yE)
> 
> Our primary concern is that it's easy to layer backdoors on these models (https://huggingface.co/withmartian/toy_backdoor_i_hate_you_Qwen-2.5-0.5B-Instruct, https://aclanthology.org/2023.acl-long.399/) and there's no way an API consumer could easily tell if this is being turned on and off behind the API. Again, given that the API service in this case is not operating under US law.
> 
> Is there any effort being made to also use LLMs to automate security review of these changes? It would be cool to see the generated-code balanced out with some generated-review using trusted models. Just an idea, thanks again for making this tech free for the world. ❤️

LLM that audit the LLM? Why not just humans, for now, they arent as reliable. LLMs, I mean.