@netrunnereve 2025-01-28
I'll add a comment here as someone who worked on quite a bit of our (AVX) vec_dot implementations...

The actual porting to SIMD or GPU code (which is also SIMD in a way) doesn't take the most time, but rather it's the low level optimization that's a headache. There you're studying the assembly, looking through the manufacturer tables to see instruction latencies, drawing out the pipeline diagrams to keep the whole CPU occupied, counting registers, and so forth. There are also tradeoffs with register use versus execution unit utilization and so on, and a lot of those things depend on the architecture.

For the original AVX I target Sandy and Ivy Bridge since that's what everyone runs it on, and I'm able to focus specifically on getting it fast there. I expect things to be messier with AVX2 with both AMD and Intel in the picture, and when I'm working on Vulkan with a huge range of different GPUs I've seen that a change that makes one architecture way faster can cause major regressions with another one.

Then again, even a bad SIMD implementation is often faster than scalar so... yeah. But unless the LLM is able to go through the full optimization cycle this won't beat a human engineer.